RDoc-2464_LongCountAsync - separated UseCount.cs and  HowToCountResults.cs to avoid ambiguity
