+++ Documentation/5.2/Samples/csharp/Raven.Documentation.Samples/ClientApi/Session/Querying/HowToCountResults.cs
@@ -7,42 +7,11 @@ using Raven.Documentation.Samples.Orders;
 using static NodaTime.TimeZones.ZoneEqualityComparer;
 using Xunit;
 using System.Threading.Tasks;
-using System.Linq;
 
 namespace Raven.Documentation.Samples.ClientApi.Session.Querying
 {
     public class CountQueryResults
     {
-        public void CanUseCount(Options options)
-        {
-            using (var store = new DocumentStore())
-            {
-                using (var s = store.OpenSession())
-                {
-                    s.Store(new User
-                    {
-                        Name = "John"
-                    });
-
-                    s.SaveChanges();
-                }
-
-                using (var s = store.OpenSession())
-                {
-                    QueryStatistics stats;
-
-                    #region Count
-                    // Use Count in a synchronous session
-                    System.Int32 Count = 
-                        s.Query<User>()
-                            .Statistics(out stats)
-                            .Search(u => u.Name, "John")
-                            .Count();
-                    #endregion
-                }
-            }
-        }
-
         public async Task CanUseCountAsync(Options options)
         {
             using (var store = new DocumentStore())
